move: query, dto files
