+++ video-service/src/video/application/query/get-video-detail.query.ts
