--- video-service/src/video/interface/dto/get-thumb-nail-path.dto.ts
@@ -1,4 +1,0 @@
-export class GetThumbNailPathDto {
-    videoId: string;
-    thumbNailPath: string
-}
