--- video-service/src/video/query/get-thumb-nail-path.query.ts
@@ -1,5 +1,0 @@
-import { IQuery } from "@nestjs/cqrs";
-
-export class GetThumbNailPathQuery implements IQuery {
-    constructor( public readonly userId: string ) {}
-}
